meta-mender-toradex-nxp: initial template files
Toradex does not use 'poky' in upstream configuration,
which results in that this vendor layer will be a bit different
from the norm.

Toradex provides there own wrapper script to setup the environment
called 'export', we will override this file with a custom one
that can be found in meta-mender-toradex-nxp to make sure our
"appends" are used.

We do also provide templates for a more minimalistic configuration,
which uses poky and only fetching meta-toradex-nxp BSP layer.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>