meta-mender-toradex-nxp: initial template files
Toradex does not use 'poky' in upstream configuration, and instead
provide a distrobution based on Angstrom + oe-core. To simplify
the setup a bit the template files here are based on poky and more
aligned with the rest of boards and provides a good enough base
to evaluate Mender on Toradex NXP SoMs.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>