WIP: bbclass to create dir-overlay Mender Artifacts during a Yocto build
This enables to create overlay directory trees by including the following:

     inherit dir-overlay-rootfs

in a recipes, which would ensure that the content of that recipe would be
included in the overlay directory tree. The dir-overlay Mender Artifact will
be in the deploy directory with a ".dir.overlay.mender" suffix and the artifact
name will be set to "${MENDER_ARTIFACT_NAME}-dir-overlay", which enables one to
utilize one build to generate Mender Artifacts for both image based and using
the dir-overlay module

This enables creation for example
application updaters, even if the "application" is split up in many recipes.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>